api/requirements.txt
pydantic
websockets
python-dotenv
azure-identity
fastapi[standard]
azure-ai-inference
opentelemetry-instrumentation
azure-monitor-opentelemetry-exporter
opentelemetry-instrumentation-fastapi
openai[realtime]
azure-cosmos
aiohttp==3.11.18
aiofiles
types-aiofiles
mypy
azure-ai-projects==1.0.0b10
azure-identity
prompty[azure]==1.0.0a1
azure-storage-blob
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
aioresponses>=0.7.4
tweepy
requests
